+++ src/next-app/components/Button.tsx
@@ -1,4 +1,4 @@
-import styles from "../styles/Button.module.css";
+import styles from "../styles/components/Button.module.css";
 
 type ButtonProps = {
   title: string;
