+++ src/next-app/components/Header.tsx
@@ -1,4 +1,4 @@
-import styles from "../styles/Header.module.css";
+import styles from "../styles/components/Header.module.css";
 import BackButton from "./BackButton";
 
 type HeaderProps = {
