+++ src/next-app/components/Link.tsx
@@ -1,4 +1,4 @@
-import styles from "../styles/Link.module.css";
+import styles from "../styles/components/Link.module.css";
 
 type LinkProps = {
   text: string;
