+++ src/next-app/components/Modal.tsx
@@ -1,5 +1,5 @@
 import { ReactElement, useEffect, useState } from "react";
-import styles from "../styles/Modal.module.css";
+import styles from "../styles/components/Modal.module.css";
 import { GrClose } from "react-icons/gr";
 
 type ModalProps = {
