+++ src/next-app/pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Button from "../components/Button";
-import styles from "../styles/Home.module.css";
+import styles from "../styles/pages/Home.module.css";
 
 const Home: NextPage = () => {
   return (
